requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-groq==0.1.0
pymongo==4.5.0
python-dotenv==1.0.0